New changes

+++ database/migrations/2023_01_06_003_create_purchase_stocks_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
         Schema::create('purchase_stocks', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('length');
-            $table->integer('width');
-            $table->integer('thikness');
+            $table->decimal('length',10,2); 
+            $table->decimal('width',10,2);
+            $table->decimal('thikness',10,2);
             $table->integer('quantity');
             $table->timestamps();
         });
